requirements.txt
pandas

django==3.2.5
djangorestframework==3.12.4
drf-yasg
celery==4.4.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
flower==0.9.5

